+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * Enhancements to GandiV5::Email::Mailbox::Responder
   * Add #enable(message:, ends_at:, starts_at: Time.now) to enable the auto responder in Gandi
   * Add #disable to disable the auto responder in Gandi
+* Enhancements to GandiV5::Email::Slot
+  * .create now returns created slot
+  * #delete now checks for inactiveness and refundableness
 * Add support for ruby 2.6.3
 
 ## Version 0.1.0